book service and controller functions modified

+++ src/main/java/com/rest_api/fs14backend/book/BookController.java
@@ -4,9 +4,13 @@ package com.rest_api.fs14backend.book;
 import com.rest_api.fs14backend.author.Author;
 import com.rest_api.fs14backend.author.AuthorService;
 import com.rest_api.fs14backend.category.Category;
+import com.rest_api.fs14backend.category.CategoryDTO;
 import com.rest_api.fs14backend.category.CategoryService;
 
+import com.rest_api.fs14backend.loan.Loan;
+import com.rest_api.fs14backend.loan.LoanService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -28,60 +32,66 @@ public class BookController {
   private AuthorService authorService;
   
   @Autowired
+  private LoanService loanService;
+  
+  @Autowired
   private BookMapper bookMapper;
 
   
 
   @GetMapping("/")
-  public List<Book> getBooks() {
-    return bookService.getAllBooks();
+  public ResponseEntity<List<Book>> getBooks() {
+    return ResponseEntity.ok(bookService.getAllBooks());
   }
 
   @GetMapping(value = "/{isbn}")
   public Optional<Book> getBookByIsbn(@PathVariable Long isbn) {
     return bookService.findById(isbn);
   }
-
+ 
+  
   @DeleteMapping(value = "/{isbn}")
   public ResponseEntity<?> deleteBook(@PathVariable Long isbn) throws Exception {
-    return bookService.deleteBook(isbn);
+    Optional<Book> bookToDelete=bookService.findById(isbn);
+    if(bookToDelete.isPresent()){
+      Loan bookIsInLoan=loanService.ifBookIsInLoan(bookToDelete.get());
+      if(bookIsInLoan!=null){
+        return ResponseEntity.badRequest().body("Book is in loan!");
+      }else{
+        boolean deletedBook=bookService.deleteBook(isbn);
+        if(deletedBook){
+          return ResponseEntity.ok(isbn);
+        }
+     }
+    }
+    return ResponseEntity.badRequest().body("Book with isbn " + isbn + " not found!");
+    
   }
 
-  /*@PostMapping
-  public void createOne(@RequestBody Book book) {
-    bookService.addOneBook(book);
-  }*/
   
-  //creating a new book
   @PostMapping("/")
-  public ResponseEntity<?> createOne(@RequestBody BookDTO bookDTO) {
-    UUID categoryId = bookDTO.getCategoryId();
-    List<UUID> authorIdList = bookDTO.getAuthorIdList();
-    Category category = categoryService.findById(categoryId);
-    //List<B> bList = aList.stream().map(A::getB).collect(Collectors.toList());
-    List<Author> authorList = new ArrayList<>();
-    authorIdList.forEach(a -> authorList.add(authorService.findById(a)));
+  public ResponseEntity<?> createOne(@RequestBody BookDTO bookDTO) throws Exception{
+    Optional<Book> bookToAdd = bookService.findById(bookDTO.getISBN());
     
-    Book book = bookMapper.newBook(bookDTO, category, authorList);
+    if(bookToAdd.isPresent()){
+      return ResponseEntity.status(HttpStatus.CONFLICT).body("Book with isbn " + bookDTO.getISBN() + " already exists!");
+    }else{
+      UUID categoryId = bookDTO.getCategoryId();
+      List<UUID> authorIdList = bookDTO.getAuthorIdList();
+      Category category = categoryService.findById(categoryId);
+      List<Author> authorList = new ArrayList<>();
+      authorIdList.forEach(a -> authorList.add(authorService.findById(a)));
+      Book book = bookMapper.newBook(bookDTO, category, authorList);
+      Book savedBook=bookService.addBook(book);
+      if(savedBook!=null){
+        return ResponseEntity.ok(savedBook);
+      }else{
+        return ResponseEntity.badRequest().body("Failed to create book");
+      }
+    }
     
-    return bookService.createOne(book);
   }
-  
-  //json value for post method
-  /*{
-    "categoryId":"81d1b149-0f37-4c21-aca8-69d6975aee43",
-     "isbn":"1234",
-     "title":"A Haunted House",
-     "publishedDate":"2000-10-31",
-     "description":"Very creepy",
-     "status":"AVAILABLE",
-     "publishers":"QP"
-  }*/
-  
-  /*@PutMapping(value = "/{isbn}")
-  public void updateBook(@PathVariable Long isbn, @RequestBody Book book) throws Exception {
-    bookService.updateBook(isbn, book);
-  }*/
+
   
   @PutMapping(value = "/{isbn}")
   public ResponseEntity<?> updateBook(@PathVariable Long isbn, @RequestBody BookDTO bookDTO) throws Exception {
@@ -92,6 +102,11 @@ public class BookController {
     authorIdList.forEach(a -> authorList.add(authorService.findById(a)));
     
     Book book = bookMapper.newBook(bookDTO, category, authorList);
-    return bookService.updateBook(isbn, book);
+    Book updatedBook= bookService.updateBook(isbn, book);
+    if(updatedBook!=null){
+      return ResponseEntity.ok(updatedBook);
+    }else{
+      return ResponseEntity.badRequest().body("Book with isbn " + isbn + " not Found!" );
+    }
   }
 }
